backend/requirements.txt
Flask
python-dotenv
flask-cors
SQLAlchemy
psycopg2-binary
pymysql
requests
python-dotenv
groq